fixed folders structure
