+++ src/main/java/com/example/ticketReservationSystem/flight/Flight.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.flight;
+
+public class Flight {
+}
