+++ src/main/java/com/example/ticketReservationSystem/ticket/TicketController.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.ticket;
+
+public class TicketController {
+}
