+++ src/main/java/com/example/ticketReservationSystem/user/User.java
@@ -1,36 +1,35 @@
-package user;
+package com.example.ticketReservationSystem.user;
 
-import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ticket.Ticket;
+import com.example.ticketReservationSystem.ticket.Ticket;
 
 import javax.persistence.*;
 import java.util.List;
 
-@Builder
-@Data
-@Table
 @Entity
+@Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
+@Table
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @NotNull
+
     private String name;
-    @NotNull
+
     private String surname;
-    @NotNull
+
     private String email;
-    @NotNull
+
     private String password;
-//    @OneToMany
-//    List<Ticket> tickets;
-//
+    @OneToMany
+    List<Ticket> tickets;
+
     
 
 }
