+++ src/main/java/com/example/ticketReservationSystem/user/UserController.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.user;
+
+public class UserController {
+}
