+++ src/main/java/com/example/ticketReservationSystem/user/UserRepository.java
@@ -1,7 +1,7 @@
-package user;
+package com.example.ticketReservationSystem.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ticket.Ticket;
+import com.example.ticketReservationSystem.ticket.Ticket;
 
 public interface UserRepository extends JpaRepository<Ticket, Long> {
 }
