+++ src/main/java/com/example/ticketReservationSystem/user/UserService.java
@@ -1,0 +1,4 @@
+package com.example.ticketReservationSystem.user;
+
+public class UserService {
+}
