--- src/main/java/flight/Flight.java
@@ -1,4 +1,0 @@
-package flight;
-
-public class Flight {
-}
