--- src/main/java/flight/FlightController.java
@@ -1,4 +1,0 @@
-package flight;
-
-public class FlightController {
-}
