--- src/main/java/flight/FlightRepository.java
@@ -1,4 +1,0 @@
-package flight;
-
-public interface FlightRepository {
-}
