--- src/main/java/flight/FlightService.java
@@ -1,4 +1,0 @@
-package flight;
-
-public class FlightService {
-}
