--- src/main/java/ticket/TicketController.java
@@ -1,4 +1,0 @@
-package ticket;
-
-public class TicketController {
-}
