--- src/main/java/ticket/TicketService.java
@@ -1,4 +1,0 @@
-package ticket;
-
-public class TicketService {
-}
