--- src/main/java/user/UserController.java
@@ -1,4 +1,0 @@
-package user;
-
-public class UserController {
-}
