--- src/main/java/user/UserService.java
@@ -1,4 +1,0 @@
-package user;
-
-public class UserService {
-}
