--- src/main/java/userType/UserType.java
@@ -1,4 +1,0 @@
-package userType;
-
-public class UserType {
-}
