+++ src/main/resources/application.properties
@@ -2,5 +2,7 @@ spring.datasource.url=jdbc:mysql://localhost:3306/reservation?useSSL=false
 spring.h2.console.enabled=true
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+
 spring.datasource.username=root
 spring.datasource.password=root
